Reorganize RabbitPEPS configuration: move class to new package and restore implementation
